+++ changelog.txt
@@ -2,8 +2,9 @@ WEIXIN-POPULAR CHANGELOG
 ===========================
 https://github.com/liyiorg/weixin-popular
 
-Changes in version 2.8.16 (2017-12-?)
+Changes in version 2.8.16 (2018-01-06)
 -------------------------------------
+* #129 GethbinfoResult hblist xml字段解析错误
 * #124 Jscode2sessionResult 中添加unionid
 * #121 QrcodeAPI 添加带字符串参数的临时二维码生成
 *
