+++ pom.xml
@@ -4,7 +4,7 @@
 	
 	<groupId>com.github.liyiorg</groupId>
 	<artifactId>weixin-popular</artifactId>
-	<version>2.8.15</version>
+	<version>2.8.16</version>
 	
 	<name>weixin-popular</name>
     <description>The weixin-popular is a JAVA SDK for weixin. Weixin web url is https://mp.weixin.qq.com.</description>
