+++ src/main/java/weixin/popular/Version.java
@@ -2,5 +2,5 @@ package weixin.popular;
 
 public interface Version {
 	
-	String VERSION = "2.8.15";
+	String VERSION = "2.8.16";
 }
