+++ src/main/java/weixin/popular/bean/paymch/GethbinfoResult.java
@@ -1,7 +1,11 @@
 package weixin.popular.bean.paymch;
 
+import java.util.List;
+
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
+import javax.xml.bind.annotation.XmlElement;
+import javax.xml.bind.annotation.XmlElementWrapper;
 import javax.xml.bind.annotation.XmlRootElement;
 
 @XmlRootElement(name = "xml")
@@ -34,10 +38,9 @@ public class GethbinfoResult extends MchBase{
 	private String wishing;
 	private String remark;
 	private String act_name;
-	private String hblist;
-	private String openid;
-	private Integer amount;
-	private String rcv_time;
+	@XmlElementWrapper(name="hblist")
+	@XmlElement(name="hbinfo")
+	private List<Hbinfo> hblist;
 
 	public String getMch_billno() {
 		return mch_billno;
@@ -151,36 +154,12 @@ public class GethbinfoResult extends MchBase{
 		this.act_name = act_name;
 	}
 
-	public String getHblist() {
+	public List<Hbinfo> getHblist() {
 		return hblist;
 	}
 
-	public void setHblist(String hblist) {
+	public void setHblist(List<Hbinfo> hblist) {
 		this.hblist = hblist;
-	}
-
-	public String getOpenid() {
-		return openid;
-	}
-
-	public void setOpenid(String openid) {
-		this.openid = openid;
-	}
-
-	public Integer getAmount() {
-		return amount;
-	}
-
-	public void setAmount(Integer amount) {
-		this.amount = amount;
-	}
-
-	public String getRcv_time() {
-		return rcv_time;
-	}
-
-	public void setRcv_time(String rcv_time) {
-		this.rcv_time = rcv_time;
 	}
 	
 }
