+++ src/main/java/weixin/popular/bean/paymch/Hbinfo.java
@@ -1,0 +1,35 @@
+package weixin.popular.bean.paymch;
+
+public class Hbinfo {
+
+	private String openid;
+
+	private Integer amount;
+
+	private String rcv_time;
+
+	public String getOpenid() {
+		return openid;
+	}
+
+	public void setOpenid(String openid) {
+		this.openid = openid;
+	}
+
+	public Integer getAmount() {
+		return amount;
+	}
+
+	public void setAmount(Integer amount) {
+		this.amount = amount;
+	}
+
+	public String getRcv_time() {
+		return rcv_time;
+	}
+
+	public void setRcv_time(String rcv_time) {
+		this.rcv_time = rcv_time;
+	}
+
+}
